components/backend/requirements.txt
fastapi==0.113.0
uvicorn==0.30.6
pydantic-settings==2.4.0
sqlalchemy==2.0.34
alembic==1.13.2
python-jose==3.3.0
attrs==24.2.0
python-json-logger==2.0.7
asyncpg==0.29.0
tinkoff-investments == 0.2.0b105